+++ doc/rst/MochiKit/VersionHistory.rst
@@ -1,5 +1,6 @@
 2006-XX-XX      v1.4
 
+- MochiKit.DOM.formContents now handles option value="" correctly.
 - MochiKit.DOM now checks for undefined className.
 - MochiKit.Iter.groupby() now uses compare() to group rather than == and !=
 - serializeJSON no longer serializes undefined, as per the JSON spec